+++ flowmachine/flowmachine/core/cache.py
@@ -31,7 +31,7 @@ if TYPE_CHECKING:
 
 import structlog
 
-logger = structlog.get_logger(__name__)
+logger = structlog.get_logger("flowmachine.debug", submodule=__name__)
 
 
 def write_query_to_cache(